gitfam/pyproject.toml
[project]
name = "gitfam"
version = "0.1.0"
description = "CLI tools for managing family history projects"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "GitFam Contributors"}
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "questionary>=2.0.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "pygments>=2.17.0",
]

[project.scripts]
gitfam = "gitfam.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]
